requirements.txt
requests
python-dotenv
anthropic
langchain
langchain-anthropic
